part4-ML/from_1021_Django/django_crud/requirements.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
django-extensions==2.2.5
Faker==2.0.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
